Compiler: Fixed #CompileUMinusNum, where it was casting Float to float instead of to double. Added project for unit testing. Change method binder's #Bind signature to have a CallInfo instead of a CallSite. Added ruby method Fixnum#bit_length, based on the C# methods Fixnum#BitLength and Fixnum#LeadingZeros. Fixed Object#Send.

+++ Mint.Compiler/Compilation/Compiler.cs
@@ -227,7 +227,7 @@ namespace Mint.Compilation
             }
             else if(number is Float)
             {
-                number = new Float(-(float) (Float) number);
+                number = new Float(-(double) (Float) number);
             }
             else if(number is Complex)
             {
